requirements.txt
requests~=2.31.0
python-dotenv==1.0.0
Brotli==1.1.0
APScheduler==3.10.4
beautifulsoup4==4.12.2